refactor: Move styles from organism
